DNS_Resolver/requirements.txt
dnslib
python-dotenv